UPD | protection against ddos #2

+++ include/worker/ManapiBaseWorker.hpp
@@ -106,7 +106,7 @@ namespace manapi::net::worker {
             ssize_t transfered_k;
         };
 
-        typedef std::map<uint16_t, shared_conn> conn_by_port;
+        typedef std::map<uintptr_t, shared_conn> conn_by_port;
 
         typedef std::map<std::array<char, 17>, conn_by_port> conns_by_ip;
